services/content-quiz-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
asyncpg==0.29.0
aiosqlite==0.19.0
httpx==0.25.0
openai==1.3.0
PyPDF2==3.0.1
python-magic==0.4.27
pillow==10.0.0
cloudinary==1.36.0
boto3==1.28.0
python-jose[cryptography]==3.3.0
requests==2.31.0

redis>=4.5.0
bleach>=6.0.0
email-validator>=1.3.0
phonenumbers>=8.13.0
